🚚 moved functions to utils file (similar to MQT Predictor where the same functions are stored in the same place)
